♻️ Refactor : 코드 리팩토링

+++ src/main/java/com/example/bookrecordService/domain/Board/Repository/BoardRepository.java
@@ -2,6 +2,7 @@ package com.example.bookrecordService.domain.Board.Repository;
 
 import com.example.bookrecordService.domain.Board.Entity.Board;
 import com.example.bookrecordService.domain.Comment.Entity.Comment;
+import com.example.bookrecordService.global.exception.ExceptionType;
 import com.example.bookrecordService.global.exception.NotFoundByIdException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -9,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
     default Board findByIdOrElseThrow(Long id) {
         return findById(id).orElseThrow(() -> new NotFoundByIdException(ExceptionType.BOARD_NOT_FOUND));
     }
